requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
whitenoise==6.6.0
